requirements.txt
aiofiles~=24.1.0
colorlog~=6.9.0
python-dotenv~=1.1.0
psutil~=7.0.0